refactor config to diferenciate specific web over others
