+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/FilePersistenceConfig.kt
@@ -19,10 +19,15 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
 import uk.ac.ebi.extended.serialization.service.ExtSerializationService
 import uk.ac.ebi.fire.client.integration.web.FireClient
+import uk.ac.ebi.fire.client.integration.web.FireClientFactory
+import uk.ac.ebi.fire.client.integration.web.FireConfig
+import uk.ac.ebi.fire.client.integration.web.RetryConfig
+import uk.ac.ebi.fire.client.integration.web.S3Config
 import java.io.File
+import kotlin.time.Duration.Companion.minutes
 
 @Configuration
-@Import(value = [SqlPersistenceConfig::class, FileSystemConfig::class])
+@Import(value = [SqlPersistenceConfig::class, GeneralConfig::class])
 class FilePersistenceConfig(
     private val folderResolver: SubmissionFolderResolver,
     private val properties: ApplicationProperties,
@@ -62,4 +67,31 @@ class FilePersistenceConfig(
         toSubmissionMapper: ToSubmissionMapper,
         toFileListMapper: ToFileListMapper,
     ): PageTabUtil = PageTabUtil(serializationService, toSubmissionMapper, toFileListMapper)
+
+    @Bean
+    fun fireClient(properties: ApplicationProperties): FireClient {
+        val fireProps = properties.fire
+        val retryProps = fireProps.retry
+        return FireClientFactory.create(
+            FireConfig(
+                fireHost = fireProps.host,
+                fireVersion = fireProps.version,
+                username = fireProps.username,
+                password = properties.fire.password
+            ),
+            S3Config(
+                accessKey = fireProps.s3.accessKey,
+                secretKey = fireProps.s3.secretKey,
+                region = fireProps.s3.region,
+                endpoint = fireProps.s3.endpoint,
+                bucket = fireProps.s3.bucket
+            ),
+            RetryConfig(
+                maxAttempts = retryProps.maxAttempts,
+                initialInterval = retryProps.initialInterval,
+                multiplier = retryProps.multiplier,
+                maxInterval = retryProps.maxInterval.minutes.inWholeMilliseconds,
+            )
+        )
+    }
 }
