+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/GeneralConfig.kt
@@ -2,7 +2,7 @@ package ac.uk.ebi.biostd.common.config
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
-import ac.uk.ebi.biostd.submission.helpers.TempFileGenerator
+import ac.uk.ebi.biostd.stats.web.TempFileGenerator
 import ac.uk.ebi.biostd.submission.service.FileSourcesService
 import ebi.ac.uk.ftp.FtpClient
 import org.springframework.boot.context.properties.EnableConfigurationProperties
@@ -11,6 +11,8 @@ import org.springframework.context.annotation.Configuration
 import uk.ac.ebi.fire.client.integration.web.FireClient
 import uk.ac.ebi.io.builder.FilesSourceListBuilder
 import uk.ac.ebi.io.config.FilesSourceConfig
+import uk.ac.ebi.serialization.common.FilesResolver
+import java.io.File
 import java.nio.file.Paths
 
 @Configuration
@@ -34,4 +36,8 @@ internal class GeneralConfig {
 
     @Bean
     fun fileSourcesService(builder: FilesSourceListBuilder): FileSourcesService = FileSourcesService(builder)
+
+    @Bean
+    fun extFilesResolver(properties: ApplicationProperties): FilesResolver =
+        FilesResolver(File(properties.requestFilesPath))
 }
