+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/JmsPublishingConfig.kt
@@ -2,14 +2,18 @@ package ac.uk.ebi.biostd.common.config
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
 import org.springframework.amqp.rabbit.connection.ConnectionFactory
+import org.springframework.amqp.rabbit.core.RabbitTemplate
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter
+import org.springframework.amqp.support.converter.MessageConverter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import uk.ac.ebi.events.config.EventsProperties
 import uk.ac.ebi.events.config.EventsPublisherConfig
 import uk.ac.ebi.events.service.EventsPublisherService
+import uk.ac.ebi.extended.serialization.service.ExtSerializationService
 
 @Configuration
-class EventsConfig(
+class JmsPublishingConfig(
     private val applicationProperties: ApplicationProperties,
 ) {
     @Bean
@@ -27,4 +31,14 @@ class EventsConfig(
     fun eventsPublisherService(
         eventsPublisherConfig: EventsPublisherConfig,
     ): EventsPublisherService = eventsPublisherConfig.eventsPublisherService()
+
+    @Bean
+    fun myRabbitTemplate(connectionFactory: ConnectionFactory): RabbitTemplate {
+        val rabbitTemplate = RabbitTemplate(connectionFactory)
+        rabbitTemplate.messageConverter = messageConverter()
+        return rabbitTemplate
+    }
+
+    @Bean
+    fun messageConverter(): MessageConverter = Jackson2JsonMessageConverter(ExtSerializationService.mapper)
 }
