+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/domain/service/SubmissionMessageListener.kt
@@ -1,6 +1,6 @@
 package ac.uk.ebi.biostd.submission.domain.service
 
-import ac.uk.ebi.biostd.common.config.LISTENER_FACTORY_NAME
+import ac.uk.ebi.biostd.common.config.internal.LISTENER_FACTORY_NAME
 import ac.uk.ebi.biostd.submission.submitter.SubmissionStagesHandler
 import ebi.ac.uk.extended.events.RequestCheckedReleased
 import ebi.ac.uk.extended.events.RequestCleaned
@@ -22,41 +22,57 @@ class SubmissionMessageListener(
 ) {
     @RabbitHandler
     fun indexRequest(rqt: RequestCreated) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received index message for submission $accNo, version: $version" }
         stagesHandler.indexRequest(rqt)
     }
 
     @RabbitHandler
     fun loadRequest(rqt: RequestIndexed) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received load message for submission $accNo, version: $version" }
         stagesHandler.loadRequest(rqt)
     }
 
     @RabbitHandler
     fun cleanRequest(rqt: RequestLoaded) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received clean message for submission $accNo, version: $version" }
         stagesHandler.cleanRequest(rqt)
     }
 
     @RabbitHandler
     fun copyRequestFiles(rqt: RequestCleaned) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received persist files message for submission $accNo, version: $version" }
         stagesHandler.copyRequestFiles(rqt)
     }
 
     @RabbitHandler
     fun checkReleased(rqt: RequestFilesCopied) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received check release status message for submission $accNo, version: $version" }
         stagesHandler.checkReleased(rqt)
     }
 
     @RabbitHandler
     fun saveSubmission(rqt: RequestCheckedReleased) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received save submission message for submission $accNo, version: $version" }
         stagesHandler.saveSubmission(rqt)
     }
 
     @RabbitHandler
     fun finalizeRequest(rqt: RequestPersisted) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received finalize submission message for submission $accNo, version: $version" }
         stagesHandler.finalizeRequest(rqt)
     }
 
     @RabbitHandler
     fun calculateStats(rqt: RequestFinalized) {
+        val (accNo, version) = rqt
+        logger.info { "$accNo, Received calculate status message for submission $accNo, version: $version" }
         stagesHandler.calculateStats(rqt)
     }
 }
